+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Added
+- Active scan rule "PTK Scan Rules" that runs all of the PTK scan rules enabled.
+
 ## 0.4.0
 ### Added
 - Keep browsers open while scanning in automation mode